fix: refactor signin logic
