+++ src/app/(site)/beaches/[slug]/page.tsx
@@ -2,7 +2,7 @@ import { PrismaClient } from '@prisma/client';
 import Image from 'next/image';
 import Link from 'next/link';
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/utils/auth";
 
 const prisma = new PrismaClient();
