--- src/app/api/auth/signin/route.ts
@@ -1,88 +1,0 @@
-import NextAuth, { AuthOptions } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-import bcrypt from "bcryptjs";
-import { JWT } from "next-auth/jwt";
-import { PrismaClient, User as PrismaUser } from "@prisma/client";
-
-const prisma = new PrismaClient();
-
-interface ExtendedUser extends PrismaUser {
-  isAdmin: boolean;
-  isPaidSubscriber: boolean;
-}
-
-// Extend JWT type
-interface ExtendedJWT extends JWT {
-  id: string;
-  email: string;
-  isAdmin: boolean;
-}
-
-export const authOptions: AuthOptions = {
-  session: {
-    strategy: "jwt",
-  },
-  providers: [
-    CredentialsProvider({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "email" },
-        password: { label: "Password", type: "password" },
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) {
-          throw new Error("Missing email or password");
-        }
-
-        // Find user by email
-        const user = await prisma.user.findUnique({
-          where: { email: credentials.email },
-        });
-
-        if (!user || !user.password) {
-          throw new Error("Invalid email or password");
-        }
-
-        // Check password
-        const isValidPassword = await bcrypt.compare(credentials.password, user.password);
-        if (!isValidPassword) {
-          throw new Error("Invalid password");
-        }
-
-        console.log("User being returned:", { id: user.id, email: user.email, isAdmin: user.isAdmin, isPaidSubscriber: user.isPaidSubscriber});
-
-        return { id: user.id, email: user.email, isAdmin: user.isAdmin, isPaidSubscriber: user.isPaidSubscriber };
-      },
-    }),
-  ],
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        const extendedUser = user as ExtendedUser;
-        token.id = extendedUser.id;
-        token.email = extendedUser.email;
-        token.isAdmin = extendedUser.isAdmin;
-        token.isPaidSubscriber = extendedUser.isPaidSubscriber;
-        console.log("JWT token after login:", token);
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (session.user) {
-        session.user.id = token.id as string;
-        session.user.email = token.email as string;
-        session.user.isAdmin = token.isAdmin as boolean;
-        session.user.isPaidSubscriber = token.isPaidSubscriber as boolean;
-        console.log("Session object:", session);
-      }
-      return session;
-    },
-  },
-  
-  pages: {
-    signIn: "/signin",
-  },
-};
-
-const handler = NextAuth(authOptions);
-export { handler as GET, handler as POST };
