+++ src/app/api/user-subscription/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/utils/auth";
 import { PrismaClient } from "@prisma/client";
 
 const prisma = new PrismaClient();
