+++ src/utils/auth.ts
@@ -1,118 +1,99 @@
-import bcrypt from "bcrypt";
-import { NextAuthOptions } from "next-auth";
+import NextAuth, { NextAuthOptions, Session, User as NextAuthUser } from "next-auth";
 import CredentialsProvider from "next-auth/providers/credentials";
-import GitHubProvider from "next-auth/providers/github";
-import GoogleProvider from "next-auth/providers/google";
-import EmailProvider from "next-auth/providers/email";
-import { PrismaAdapter } from "@auth/prisma-adapter";
+import bcrypt from "bcryptjs";
 import { PrismaClient } from "@prisma/client";
-import { prisma } from "./prismaDB";
-import type { Adapter } from "next-auth/adapters";
+import { JWT } from "next-auth/jwt";
 
+// Initialize Prisma Client
+const prisma = new PrismaClient();
+
+// Extend NextAuth types (ensure this is at the top)
+declare module "next-auth" {
+  interface User {
+    id: string;
+    email: string;
+    isAdmin: boolean;
+    isPaidSubscriber: boolean;
+  }
+
+  interface Session {
+    user: User;
+  }
+
+  interface JWT {
+    id: string;
+    email: string;
+    isAdmin: boolean;
+    isPaidSubscriber: boolean;
+  }
+}
+
+// Define authentication options inline (do not export this)
 export const authOptions: NextAuthOptions = {
-  pages: {
-    signIn: "/auth/signin",
-  },
-  adapter: PrismaAdapter(prisma) as Adapter,
-  secret: process.env.SECRET,
   session: {
     strategy: "jwt",
   },
-
   providers: [
     CredentialsProvider({
-      name: "credentials",
+      name: "Credentials",
       credentials: {
-        email: { label: "Email", type: "text", placeholder: "Jhondoe" },
+        email: { label: "Email", type: "email" },
         password: { label: "Password", type: "password" },
-        username: { label: "Username", type: "text", placeholder: "Jhon Doe" },
       },
-
       async authorize(credentials) {
-        // check to see if email and password is there
         if (!credentials?.email || !credentials?.password) {
-          throw new Error("Please enter an email or password");
+          throw new Error("Missing email or password");
         }
 
-        // check to see if user already exist
+        // Find the user by email using Prisma
         const user = await prisma.user.findUnique({
-          where: {
-            email: credentials.email,
-          },
+          where: { email: credentials.email },
         });
 
-        // if user was not found
-        if (!user || !user?.password) {
+        if (!user || !user.password) {
           throw new Error("Invalid email or password");
         }
 
-        // check to see if passwords match
-        const passwordMatch = await bcrypt.compare(
-          credentials.password,
-          user.password,
-        );
-
-        // console.log(passwordMatch);
-
-        if (!passwordMatch) {
-          console.log("test", passwordMatch);
-          throw new Error("Incorrect password");
+        // Check password validity using bcrypt
+        const isValidPassword = await bcrypt.compare(credentials.password, user.password);
+        if (!isValidPassword) {
+          throw new Error("Invalid password");
         }
 
-        return user;
+        // Return user object without sensitive data
+        return {
+          id: user.id,
+          email: user.email,
+          isAdmin: user.isAdmin,
+          isPaidSubscriber: user.isPaidSubscriber
+        };
       },
-    }),
-
-    GitHubProvider({
-      clientId: process.env.GITHUB_CLIENT_ID!,
-      clientSecret: process.env.GITHUB_CLIENT_SECRET!,
-    }),
-
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-    }),
-
-    EmailProvider({
-      server: {
-        host: process.env.EMAIL_SERVER_HOST,
-        port: Number(process.env.EMAIL_SERVER_PORT),
-        auth: {
-          user: process.env.EMAIL_SERVER_USER,
-          pass: process.env.EMAIL_SERVER_PASSWORD,
-        },
-      },
-      from: process.env.EMAIL_FROM,
     }),
   ],
-
   callbacks: {
-    jwt: async (payload: any) => {
-      const { token } = payload;
-      const user = payload.user;
-
+    async jwt({ token, user }: { token: JWT; user?: NextAuthUser }) {
       if (user) {
-        return {
-          ...token,
-          id: user.id,
-        };
+        return { ...token, id: user.id, email: user.email, isAdmin: user.isAdmin, isPaidSubscriber: user.isPaidSubscriber };
       }
       return token;
     },
 
-    session: async ({ session, token }) => {
-      if (session?.user) {
-        return {
-          ...session,
-          user: {
-            ...session.user,
-            id: token?.id,
-          },
-        };
-      }
-      return session;
+    async session({ session, token }: { session: Session; token: JWT }) {
+      return {
+        ...session,
+        user: {
+          ...session.user,
+          id: token.id,
+          email: token.email,
+          isAdmin: token.isAdmin,
+          isPaidSubscriber: token.isPaidSubscriber,
+        },
+      };
     },
   },
-
-  // debug: process.env.NODE_ENV === "developement",
+  pages: {},
 };
+
+// Create the NextAuth handler and export only GET and POST
+const handler = NextAuth(authOptions);
+export { handler as GET, handler as POST };
